requirements.txt
falcon==2.0.0
gunicorn==20.0.4
google-api-python-client==1.8.0
google_auth_oauthlib==0.4.1
redis==3.4.1